requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.22.0
pydantic[email]==2.9.2
python-dotenv==1.0.0
pymongo==4.10.1
bcrypt==4.2.0
PyJWT==2.9.0
requests==2.31.0
pillow==10.0.1
python-multipart==0.0.6
